+++ RemoteSpace/SpaceApi/Models/FileElement.cs
@@ -17,7 +17,7 @@ namespace SpaceApi
         [Required]
         public string Name { get; set; }
         [Required]
-        public decimal Weight { get; set; }
+        public long Weight { get; set; }
         [Required]
         public string Owner { get; set; }
         [Required]
